static: use vite to build assets
Assets directory is 1.5MB uncompressed
475kB compressed.

We compressed the go:embed'ed filesystem so that the final binary size
doesn't increase too much.

Add scripts to create the dist hash and to diff it

CI: bump actions, add assets check workflow

Package.json: remove jquery and bootstrap/js, use smaller plotly bundle.

Add maximize/minimize plot.
