+++ .github/workflows/assets.yml
@@ -1,0 +1,18 @@
+name: Assets
+on: [push, pull_request]
+jobs:
+  check-assets:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - name: Setup node
+        uses: actions/setup-node@v4
+        with:
+          node-version-file: internal/static/.nvmrc
+      - uses: actions/setup-go@v5
+      - run: npm install
+        working-directory: internal/static
+      - run: npm run build
+        working-directory: internal/static
+      - name: Check assets zip is up to date
+        run: ./internal/static/scripts/checkzip.sh
