+++ .github/workflows/tests-others.yml
@@ -1,15 +1,15 @@
 on: [pull_request]
 name: Tests-others
 jobs:
-  test:
+  tests-macos-windows:
     strategy:
       matrix:
         os: [macos-latest, windows-latest]
     runs-on: ${{ matrix.os }}
     steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-go@v3
+      - uses: actions/checkout@v4
+      - uses: actions/setup-go@v5
         with:
-          go-version: 1.21.x
+          go-version-file: go.mod
       - name: Tests
         run: go test -race -shuffle=on ./...
