+++ internal/static/.env.development
@@ -1,0 +1,2 @@
+VITE_WEBSOCKET_URL="ws://localhost:9090/debug/statsviz/ws"
+
