+++ internal/static/.gitignore
@@ -1,0 +1,18 @@
+# Logs
+logs
+*.log
+npm-debug.log*
+node_modules
+dist/
+
+# Editor directories and files
+.vscode/*
+!.vscode/extensions.json
+.idea
+.DS_Store
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.sw?
+
