+++ internal/static/.nvmrc
@@ -1,0 +1,1 @@
+v22.7.0
