+++ internal/static/index.html
@@ -1,61 +1,111 @@
 <!DOCTYPE html>
 <html lang="en">
-
-<head>
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="./favicon.ico" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Statsviz</title>
-    <meta charset="utf-8">
-    <meta name="description" content="Statsviz interface" />
-    <link rel="stylesheet" href="./libs/css/bootstrap.min.css" />
-    <link rel="stylesheet" href="./libs/css/bootstrap-toggle.min.css">
-    <link rel="stylesheet" href="./css/statsviz.css" />
-    <link rel="stylesheet" href="./libs/css/font.awesome-6.1.2-all.min.css">
-    <script src="./libs/js/jquery-3.6.0.min.js"></script>
-    <script src="./libs/js/plotly-basic-2.12.1.min.js"></script>
-    <script src="./libs/js/plotly-cartesian-2.12.1.min.js"></script>
-    <script type="module" src="./libs/js/bootstrap.bundle.min.js"></script>
-    <script src="./libs/js/popperjs-core2"></script>
-    <script src="./libs/js/tippy.js@6"></script>
-    <script src="./libs/js/bootstrap-toggle.min.js"></script>
-    <script src="./libs/js/font.awesome-6.1.2-all.min.js"></script>
-</head>
+  </head>
+  <body>
+    <div id="app">
+      <nav id="navbar" class="navbar navbar-expand-lg bg-body-tertiary">
+        <div class="container-fluid">
+          <a class="navbar-brand">Statsviz</a>
+          <button
+            class="navbar-toggler"
+            type="button"
+            data-bs-toggle="collapse"
+            data-bs-target="#navbarSupportedContent"
+            aria-controls="navbarSupportedContent"
+            aria-expanded="false"
+            aria-label="Toggle navigation"
+          >
+            <span class="navbar-toggler-icon"></span>
+          </button>
+          <div class="collapse navbar-collapse" id="navbarSupportedContent">
+            <ul class="navbar-nav mb-2 mb-lg-0 ms-auto">
+              <li class="nav-item">
+                <a
+                  class="bi bi-github nav-link"
+                  href="https://github.com/arl/statsviz"
+                  target="_blank"
+                >
+                  GitHub</a
+                >
+              </li>
+              <li class="nav-item">
+                <a
+                  class="nav-link"
+                  href="https://pkg.go.dev/runtime/metrics#hdr-Supported_metrics"
+                  target="_blank"
+                  >runtime/metrics</a
+                >
+              </li>
+              <li class="nav-item">
+                <select
+                  class="form-select"
+                  id="select-timerange"
+                  style="width: 120px"
+                >
+                  <option value="60">1 minute</option>
+                  <option value="300">5 minutes</option>
+                  <option value="600">10 minutes</option>
+                </select>
+              </li>
+              <li
+                class="nav-item d-flex align-items-center pt-3 pt-lg-1 ms-0 ms-lg-3"
+              >
+                <div class="form-check form-switch">
+                  <input
+                    class="form-check-input"
+                    type="checkbox"
+                    id="dark_mode_switch"
+                    name="darkmode"
+                  />
+                  <label class="form-check-label" for="darkmode"
+                    >Dark Mode</label
+                  >
+                </div>
+              </li>
 
-<body>
-    <nav id="navbar" class="navbar navbar-expand-lg navbar-light bg-light">
-        <div class="container-xxl">
-            <a class="navbar-brand" href="#">Statsviz</a>
-            <button class="navbar-toggler" type="button">
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse">
-                <ul class="navbar-nav me-auto mb-2 mb-lg-0">
-                    <li class="nav-item">
-                        <a class="nav-link" href="https://github.com/arl/statsviz">Github</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="https://pkg.go.dev/runtime/metrics#hdr-Supported_metrics">runtime/metrics</a>
-                    </li>
-                </ul>
-            </div>
-            <span title="show/hide GC events">
-                <input id="show_gc" type="checkbox" data-toggle="toggle" data-onstyle="default" data-on="<span style='color:black'><i class='fa-solid fa-broom'></i></span>" data-off="<span style='color:lightgrey'><i class='fa-solid fa-broom'></i></span>"
-            data-size="mini" data-bs-toggle="tooltip">
-            </span>
-            <select data-bs-toggle="tooltip" title="time range" id="select_timerange">
-                <option selected value="60">1 minute</option>
-                <option value="300">5 minute</option>
-                <option value="600">10 minutes</option>
-              </select>
-            <span class="action" title="play/pause">
-                <input id="play_pause" type="checkbox" data-toggle="toggle" data-onstyle="default" data-on="<i class='fa fa-play'></i>" data-off="<i class='fa fa-pause'></i>" data-size="mini" data-bs-toggle="tooltip">
-            </span>
-            <span class="action" title="color theme">
-                <input id="color_theme_sw" type="checkbox" data-toggle="toggle" data-onstyle="default" data-on="<i class='fa fa-circle-half-stroke'></i>" data-off="<i class='fa fa-circle-half-stroke'></i>" data-size="mini" data-bs-toggle="tooltip">
-            </span>
+              <li
+                class="nav-item d-flex align-items-center pt-3 pt-lg-1 ms-0 ms-lg-3"
+              >
+                <div class="form-check form-switch">
+                  <input
+                    class="form-check-input"
+                    type="checkbox"
+                    id="show_gc_switch"
+                    name="show_gc"
+                    checked
+                  />
+                  <label class="form-check-label" for="show_gc">Show GC</label>
+                </div>
+              </li>
+
+              <li
+                class="nav-item d-flex align-items-center pt-3 pt-lg-1 ms-0 ms-lg-3"
+              >
+                <div class="form-check form-switch">
+                  <input
+                    class="form-check-input"
+                    type="checkbox"
+                    id="play_pause_switch"
+                    name="play_pause"
+                    checked
+                  />
+                  <label class="form-check-label" for="play_pause"
+                    >Play/Pause</label
+                  >
+                </div>
+              </li>
+            </ul>
+          </div>
         </div>
-    </nav>
+      </nav>
 
-    <div id="plots" class="plots-wrapper"></div>
-    <script type="module" src="./js/app.js"></script>
-</body>
-
+      <div id="plots" class="plots-wrapper"></div>
+    </div>
+    <script type="module" src="/src/main.js"></script>
+  </body>
 </html>
