+++ internal/static/src/buffer.js
