+++ internal/static/src/main.js
@@ -1,0 +1,7 @@
+import "bootstrap/dist/css/bootstrap.min.css";
+import "./style.css";
+import { connect } from "./app.js";
+import * as theme from "./theme.js";
+
+connect();
+theme.updateThemeMode();
